// app/backend/optimus/types/OptimusConnectionTestRequest.ts
export type OptimusConnectionTestRequest = {
  username: string;
  api_key: string;
  action: string;
  destinatar_nume: string;
  destinatar_adresa: string;
  destinatar_localitate: string;
  destinatar_judet: string;
  destinatar_cod_postal: string;
  destinatar_contact: string;
  destinatar_telefon: string;
  colet_buc: string;
  colet_greutate: string;
  data_colectare: string;
  ramburs_valoare: string;
  ref_factura: string;
  colet_descriere: string;
  destinatar_email: string;
};